+++ EvolutionCMS.libraries.ddInstaller.php
@@ -1,45 +1,57 @@
 <?php
 /**
  * ddInstaller
- * @version 0.1.1 (2021-04-16)
+ * @version 0.2 (2024-09-13)
  * 
  * @see README.md
  * 
- * @copyright 2021 DD Group {@link https://DivanDesign.biz }
+ * @copyright 2021–2024 Ronef {@link https://Ronef.me }
  */
 
-//Simple API
+// Simple API
 class DDInstaller {
 	/**
 	 * install
-	 * @version 1.0 (2021-04-08)
+	 * @version 1.1 (2024-09-13)
 	 * 
 	 * @param $params {stdClass|arrayAssociative|stringJsonObject|stringHjsonObject|stringQueryFormatted} — @required
 	 * @param $params->url {stringUrl} — Resource GitHub URL (e. g. `https://github.com/DivanDesign/EvolutionCMS.libraries.ddTools`). @required
-	 * @param $params->type {'Snippet'|'Plugin'|'Library'} — Resource type. @required
+	 * @param [$params->type] {'Snippet'|'Plugin'|'Library'} — Resource type.
 	 * 
 	 * @return {boolean}
 	 */
 	public static function install($params){
-		//Prepare params
+		// Prepare params
 		$params = \DDTools\ObjectTools::convertType([
 			'object' => $params,
-			'type' => 'objectStdClass'
+			'type' => 'objectStdClass',
 		]);
 		
-		$installerObject = \DDInstaller\Installer::createChildInstance([
+		if (empty($params->type)){
+			$urlLowerCase = strtolower($params->url);
+			
+			// TODO: Replace `strpos` to `str_contains` (PHP >= 8.0 is required)
+			// E. g. 'EvolutionCMS.snippets.ddGetChunk'
+			if (strpos($urlLowerCase, 'snippet') !== false){
+				$params->type = 'snippet';
+			// E. g. 'EvolutionCMS.plugins.ddSendRedirect'
+			}elseif (strpos($urlLowerCase, 'plugin') !== false){
+				$params->type = 'plugin';
+			}elseif (
+				// E. g. 'EvolutionCMS.libraries.ddTools', 'EvolutionCMS.library.ddTools'
+				strpos($urlLowerCase, 'libraries') !== false
+				|| strpos($urlLowerCase, 'library') !== false
+			){
+				$params->type = 'library';
+			}
+		}
+		
+		$installerObject = \DDInstaller\Installer\Installer::createChildInstance([
 			'name' => $params->type,
-			'parentDir' =>
-				__DIR__ .
-				DIRECTORY_SEPARATOR .
-				'src' .
-				DIRECTORY_SEPARATOR .
-				'Installer'
-			,
-			//Passing parameters into constructor
+			// Passing parameters into constructor
 			'params' => [
-				'url' => $params->url
-			]
+				'url' => $params->url,
+			],
 		]);
 		
 		return $installerObject->install();