+++ src/Installer/Installer.php
@@ -1,7 +1,9 @@
 <?php
-namespace DDInstaller;
+namespace DDInstaller\Installer;
 
-abstract class Installer extends \DDTools\BaseClass {
+abstract class Installer extends \DDTools\Base\Base {
+	use \DDTools\Base\AncestorTrait;
+	
 	protected
 		/**
 		 * @property $distrData {stdClass}
@@ -14,7 +16,7 @@ abstract class Installer extends \DDTools\BaseClass {
 			'fullName' => '',
 			'shortName' => '',
 			'type' => '',
-			'owner' => ''
+			'owner' => '',
 		],
 		
 		/**
@@ -28,7 +30,7 @@ abstract class Installer extends \DDTools\BaseClass {
 			'assetsDir' => '',
 			'destinationDir' => '',
 			'cacheDir' => '',
-			'cacheFile' => ''
+			'cacheFile' => '',
 		],
 		
 		/**
@@ -38,69 +40,58 @@ abstract class Installer extends \DDTools\BaseClass {
 		 */
 		 $dbSettings = [
 			'tableName' => null,
-			'contentField' => null
+			'contentField' => null,
 		]
 	;
 	
 	/**
 	 * __construct
-	 * @version 1.0 (2021-04-08)
+	 * @version 1.0.3 (2024-09-13)
 	 * 
 	 * @param $params {stdClass|arrayAssociative|stringJsonObject|stringHjsonObject|stringQueryFormatted} — @required
 	 * @param $params->url {stringUrl} — Resource GitHub URL (e. g. `https://github.com/DivanDesign/EvolutionCMS.libraries.ddTools`). @required
 	 */
 	public function __construct($params = []){
-		//Prepare params
+		// Prepare params
 		$params = \DDTools\ObjectTools::convertType([
 			'object' => $params,
-			'type' => 'objectStdClass'
+			'type' => 'objectStdClass',
 		]);
 		
-		//Prepare field types
+		// Prepare field types
 		$this->paths = (object) $this->paths;
 		$this->distrData = (object) $this->distrData;
 		$this->dbSettings = (object) $this->dbSettings;
 		
-		//Prepare DB settings
+		// Prepare DB settings
 		if (!empty($this->dbSettings->tableName)){
 			$this->dbSettings->tableName = \ddTools::$tables[$this->dbSettings->tableName];
 		}
 		
-		//Fill distr data from URL
+		// Fill distr data from URL
 		$this->fillDistrDataFromUrl($params->url);
 		
-		//Fill distr resource type
+		// Fill distr resource type
 		$this->distrData->type =
-			//E. g. `snippet`
+			// E. g. `snippet`
 			strtolower(
-				//E. g. [`Snippet`]
-				array_slice(
-					//E. g. `['DDInstaller', 'Snippet', 'Installer']`
-					explode(
-						'\\',
-						//E. g. `DDInstaller\\Snippet\\Installer`
-						get_called_class()
-					),
-					-2,
-					1
-				)
-				//E. g. `Snippet`
-				[0]
+				// E. g. `Snippet`
+				static::getClassName()->namespaceShort
 			)
 		;
 		
-		//Fill paths
+		// Fill paths
 		$this->fillPaths();
 		
-		//Create cache dir if needed
+		// Create cache dir if needed
 		\DDTools\FilesTools::createDir([
-			'path' => $this->paths->cacheDir
+			'path' => $this->paths->cacheDir,
 		]);
 	}
 	
 	/**
 	 * fillDistrDataFromUrl
-	 * @version 1.0.1 (2021-04-16)
+	 * @version 1.0.3 (2024-09-13)
 	 * 
 	 * @desc Parses GitHub URL and fill resource data fields.
 	 * 
@@ -108,12 +99,12 @@ abstract class Installer extends \DDTools\BaseClass {
 	 */
 	protected final function fillDistrDataFromUrl($distrUrl){
 		$ownerAndRepo =
-			//E. g. `['DivanDesign', 'EvolutionCMS.libraries.ddTools']`
+			// E. g. `['DivanDesign', 'EvolutionCMS.libraries.ddTools']`
 			array_slice(
-				//E. g. `['https:', '', 'github.com', 'DivanDesign', 'EvolutionCMS.libraries.ddTools']`
+				// E. g. `['https:', '', 'github.com', 'DivanDesign', 'EvolutionCMS.libraries.ddTools']`
 				explode(
 					'/',
-					//E. g. 'https://github.com/DivanDesign/EvolutionCMS.libraries.ddTools'
+					// E. g. 'https://github.com/DivanDesign/EvolutionCMS.libraries.ddTools'
 					$distrUrl
 				),
 				-2,
@@ -124,74 +115,74 @@ abstract class Installer extends \DDTools\BaseClass {
 		$this->distrData->owner = $ownerAndRepo[0];
 		$this->distrData->fullName = $ownerAndRepo[1];
 		
-		//E. g. `['EvolutionCMS', 'libraries', 'ddTools']`
+		// E. g. `['EvolutionCMS', 'libraries', 'ddTools']`
 		$this->distrData->shortName = explode(
 			'.',
-			//E. g. `EvolutionCMS.libraries.ddTools`
+			// E. g. `EvolutionCMS.libraries.ddTools`
 			$this->distrData->fullName
 		);
-		//E. g. `ddTools`
+		// E. g. `ddTools`
 		$this->distrData->shortName = array_pop($this->distrData->shortName);
 	}
 	
 	/**
 	 * fillPaths
-	 * @version 1.0 (2021-04-03)
+	 * @version 1.0.2 (2024-09-13)
 	 * 
 	 * @return {void}
 	 */
 	protected final function fillPaths(){
-		//Path of `assets`
+		// Path of `assets`
 		$this->paths->assetsDir =
 			dirname(
 				__DIR__,
 				4
-			) .
-			DIRECTORY_SEPARATOR
+			)
+			. DIRECTORY_SEPARATOR
 		;
 		
-		//Destination path
+		// Destination path
 		$this->fillPaths_destination();
 		
-		//Cache dir
+		// Cache dir
 		$this->paths->cacheDir =
-			$this->paths->assetsDir .
-			'cache' .
-			DIRECTORY_SEPARATOR .
-			'ddInstaller' .
-			DIRECTORY_SEPARATOR
+			$this->paths->assetsDir
+			. 'cache'
+			. DIRECTORY_SEPARATOR
+			. 'ddInstaller'
+			. DIRECTORY_SEPARATOR
 		;
 		
-		//Cache file
+		// Cache file
 		$this->paths->cacheFile =
-			$this->paths->cacheDir .
-			$this->distrData->fullName .
-			'.zip'
+			$this->paths->cacheDir
+			. $this->distrData->fullName
+			. '.zip'
 		;
 	}
 	
 	/**
 	 * fillPaths_destination
-	 * @version 1.0 (2021-04-03)
+	 * @version 1.0.1 (2024-09-13)
 	 * 
 	 * @return {void}
 	 */
 	protected function fillPaths_destination(){
 		$this->paths->destinationDir =
-			$this->paths->assetsDir .
-			(
-				$this->distrData->type .
-				's'
-			) .
-			DIRECTORY_SEPARATOR .
-			$this->distrData->shortName .
-			DIRECTORY_SEPARATOR
+			$this->paths->assetsDir
+			. (
+				$this->distrData->type
+				. 's'
+			)
+			. DIRECTORY_SEPARATOR
+			. $this->distrData->shortName
+			. DIRECTORY_SEPARATOR
 		;
 	}
 	
 	/**
 	 * install
-	 * @version 1.0 (2021-04-07)
+	 * @version 1.0.2 (2024-09-13)
 	 * 
 	 * @return {boolean} — Is resource installed?
 	 */
@@ -205,74 +196,75 @@ abstract class Installer extends \DDTools\BaseClass {
 			$distrRootDir = $distrZipObject->getNameIndex(0);
 			
 			$distrComposerJson = $distrZipObject->getFromName(
-				$distrRootDir .
-				'composer.json'
+				$distrRootDir
+				. 'composer.json'
 			);
 			
 			if (
-				is_string($distrComposerJson) &&
-				!empty($distrComposerJson)
+				is_string($distrComposerJson)
+				&& !empty($distrComposerJson)
 			){
 				$distrComposerJson = \DDTools\ObjectTools::convertType([
 					'object' => $distrComposerJson,
-					'type' => 'objectStdClass'
+					'type' => 'objectStdClass',
 				]);
 			}
 			
 			if ($this->isNeedToInstall($distrComposerJson)){
-				//Just remove exist dir
+				// Just remove exist dir
 				\DDTools\FilesTools::removeDir($this->paths->destinationDir);
-				//And create again
+				// And create again
 				\DDTools\FilesTools::createDir([
-					'path' => $this->paths->destinationDir
+					'path' => $this->paths->destinationDir,
 				]);
 				
-				//Iterate over all files in the archive
+				// Iterate over all files in the archive
 				for (
-					//Skip root dir
+					// Skip root dir
 					$fileIndex = 1;
 					$fileIndex < $distrZipObject->numFiles;
 					$fileIndex++
 				){
-					//Various directory separators support
+					// Various directory separators support
 					$filePathname = str_replace(
 						'\\',
 						'/',
-						//Get current file name
+						// Get current file name
 						$distrZipObject->getNameIndex($fileIndex)
 					);
 					
-					//Remove root dir from file name
+					// Remove root dir from file name
 					$filePathname = str_replace(
 						$distrRootDir,
 						'',
 						$filePathname
 					);
 					
-					//If it is dir
+					// If it is dir
 					if (
 						substr(
 							$filePathname,
 							-1
-						) ==
-						'/'
+						)
+						== '/'
 					){
-						//Create
+						// Create
 						\DDTools\FilesTools::createDir([
 							'path' =>
-								$this->paths->destinationDir .
-								$filePathname
+								$this->paths->destinationDir
+								. $filePathname
+							,
 						]);
 					}else{
-						//If the file must be installed to DB
+						// If the file must be installed to DB
 						if (
-							$filePathname ==
-							//E. g. `ddMakeHttpRequest_snippet.php`
-							(
-								$this->distrData->shortName .
-								'_' .
-								$this->distrData->type .
-								'.php'
+							$filePathname
+							// E. g. `ddMakeHttpRequest_snippet.php`
+							== (
+								$this->distrData->shortName
+								. '_'
+								. $this->distrData->type
+								. '.php'
 							)
 						){
 							$this->installToDb([
@@ -280,12 +272,12 @@ abstract class Installer extends \DDTools\BaseClass {
 								'description' =>
 									\DDTools\ObjectTools::isPropExists([
 										'object' => $distrComposerJson,
-										'propName' => 'description'
-									]) ?
-									$distrComposerJson->description :
-									''
+										'propName' => 'description',
+									])
+									? $distrComposerJson->description
+									: ''
 								,
-								'content' => $distrZipObject->getFromIndex($fileIndex)
+								'content' => $distrZipObject->getFromIndex($fileIndex),
 							]);
 						}else{
 							file_put_contents(
@@ -312,7 +304,7 @@ abstract class Installer extends \DDTools\BaseClass {
 	
 	/**
 	 * downloadDistrZip
-	 * @version 1.0 (2021-04-08)
+	 * @version 1.0.2 (2024-09-13)
 	 * 
 	 * @return {boolean}
 	 */
@@ -323,29 +315,29 @@ abstract class Installer extends \DDTools\BaseClass {
 			'name' => 'ddMakeHttpRequest',
 			'params' => [
 				'url' =>
-					'https://api.github.com/repos/' .
-					$this->distrData->owner .
-					'/' .
-					$this->distrData->fullName .
-					'/zipball/'
+					'https://api.github.com/repos/'
+					. $this->distrData->owner
+					. '/'
+					. $this->distrData->fullName
+					. '/zipball/'
 				,
 				'userAgent' => \ddTools::$modx->getConfig('site_url'),
 				'headers' => [
-					'Accept: application/vnd.github.v3+json'
-				]
-			]
+					'Accept: application/vnd.github.v3+json',
+				],
+			],
 		]);
 		
-		//Clear all or we will get error
+		// Clear all or we will get error
 		ob_clean();
 		
-		//If we have dump
+		// If we have dump
 		if(
-			is_string($fileContent) &&
-			//If non-JSON is gotten
-			$fileContent[0] != '{'
+			is_string($fileContent)
+			// If non-JSON is gotten
+			&& $fileContent[0] != '{'
 		){
-			//Save cache file
+			// Save cache file
 			file_put_contents(
 				$this->paths->cacheFile,
 				$fileContent
@@ -359,52 +351,52 @@ abstract class Installer extends \DDTools\BaseClass {
 	
 	/**
 	 * isInstallNeeded
-	 * @version 1.0 (2021-04-07)
+	 * @version 1.0.2 (2024-09-13)
 	 * 
 	 * @param $distrComposerJson {stdClass}
 	 * 
 	 * @return {boolean}
 	 */
 	protected function isNeedToInstall($distrComposerJson){
-		//Don't want to install by default
+		// Don't want to install by default
 		$result = false;
 		
 		if (
-			//We don't do anything if the repository has no `composer.json`
-			is_object($distrComposerJson) &&
-			//Version is required
-			!empty($distrComposerJson->version)
+			// We don't do anything if the repository has no `composer.json`
+			is_object($distrComposerJson)
+			// Version is required
+			&& !empty($distrComposerJson->version)
 		){
 			$existComposerJson =
-				$this->paths->destinationDir .
-				'composer.json'
+				$this->paths->destinationDir
+				. 'composer.json'
 			;
 			
 			if (
-				//If destination composer is absent
-				!is_file($existComposerJson) ||
-				//Or invalid
-				empty($existComposerJson = file_get_contents($existComposerJson))
+				// If destination composer is absent
+				!is_file($existComposerJson)
+				// Or invalid
+				|| empty($existComposerJson = file_get_contents($existComposerJson))
 			){
-				//Just install
+				// Just install
 				$result = true;
 			}else{
 				$existComposerJson = \DDTools\ObjectTools::convertType([
 					'object' => $existComposerJson,
-					'type' => 'objectStdClass'
+					'type' => 'objectStdClass',
 				]);
 				
 				if (
-					//If destination version is absent
-					empty($existComposerJson->version) ||
-					//Or distr version > destination version
-					version_compare(
+					// If destination version is absent
+					empty($existComposerJson->version)
+					// Or distr version > destination version
+					|| version_compare(
 						$distrComposerJson->version,
 						$existComposerJson->version,
 						'>'
 					)
 				){
-					//Install
+					// Install
 					$result = true;
 				}
 			}
@@ -415,7 +407,7 @@ abstract class Installer extends \DDTools\BaseClass {
 	
 	/**
 	 * installToDb
-	 * @version 1.0.1 (2021-04-16)
+	 * @version 1.0.3 (2024-09-13)
 	 * 
 	 * @param $params {stdClass|arrayAssociative|stringJsonObject|stringHjsonObject|stringQueryFormatted} — @required
 	 * @param $params->version {string} — @required
@@ -425,10 +417,10 @@ abstract class Installer extends \DDTools\BaseClass {
 	 * @return {void}
 	 */
 	protected function installToDb($params){
-		//Prepare params
+		// Prepare params
 		$params = \DDTools\ObjectTools::convertType([
 			'object' => $params,
-			'type' => 'objectStdClass'
+			'type' => 'objectStdClass',
 		]);
 		
 		if (!empty($this->dbSettings->tableName)){
@@ -444,49 +436,46 @@ abstract class Installer extends \DDTools\BaseClass {
 			
 			$fieldsToUpdate = [
 				'description' => \ddTools::$modx->db->escape(
-					'<b>' .
-					$params->version .
-					'</b> ' .
-					$params->description
+					'<b>'
+					. $params->version
+					. '</b> '
+					. $params->description
 				),
-				$this->dbSettings->contentField => \ddTools::$modx->db->escape($params->content)
+				$this->dbSettings->contentField => \ddTools::$modx->db->escape($params->content),
 			];
 			
 			$destinationId = \ddTools::$modx->db->getValue(\ddTools::$modx->db->select(
-				//Fields
+				// Fields
 				'id',
-				//From
+				// From
 				$this->dbSettings->tableName,
-				//Where
+				// Where
 				(
-					'`name` = "' .
-					\ddTools::$modx->db->escape($this->distrData->shortName) .
-					'"'
+					'`name` = "'
+						. \ddTools::$modx->db->escape($this->distrData->shortName)
+					. '"'
 				)
 			));
 			
-			//If resource already exists
+			// If resource already exists
 			if (is_numeric($destinationId)){
 				$fieldsToUpdate['editedon'] = time();
 				
 				\ddTools::$modx->db->update(
-					//Fields
+					// Fields
 					$fieldsToUpdate,
-					//From
+					// From
 					$this->dbSettings->tableName,
-					//Where
-					(
-						'`id` = ' .
-						$destinationId
-					)
+					// Where
+					'`id` = ' . $destinationId
 				);
 			}else{
 				$fieldsToUpdate['createdon'] = time();
 				
 				\ddTools::$modx->db->insert(
-					//Fields
+					// Fields
 					$fieldsToUpdate,
-					//From
+					// From
 					$this->dbSettings->tableName
 				);
 			}