+++ src/Installer/Library/Installer.php
@@ -1,20 +1,20 @@
 <?php
-namespace DDInstaller\Library;
+namespace DDInstaller\Installer\Library;
 
-class Installer extends \DDInstaller\Installer {
+class Installer extends \DDInstaller\Installer\Installer {
 	/**
 	 * fillPaths_destination
-	 * @version 1.0 (2021-04-03)
+	 * @version 1.0.1 (2024-09-13)
 	 *
 	 * @return {void}
 	 */
 	protected final function fillPaths_destination(){
 		$this->paths->destinationDir =
-			$this->paths->assetsDir .
-			'libs' .
-			DIRECTORY_SEPARATOR .
-			$this->distrData->shortName .
-			DIRECTORY_SEPARATOR
+			$this->paths->assetsDir
+			. 'libs'
+			. DIRECTORY_SEPARATOR
+			. $this->distrData->shortName
+			. DIRECTORY_SEPARATOR
 		;
 	}
 }